Removed SANS reduction example
